backend/python-service/requirements.txt
Flask==3.1.0
Pillow==11.0.0
opencv-python-headless
numpy
pdf2image
google-cloud-vision